requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile pyproject.toml
#
contourpy==1.0.6
    # via matplotlib
cycler==0.11.0
    # via matplotlib
fonttools==4.38.0
    # via matplotlib
kiwisolver==1.4.4
    # via matplotlib
llvmlite==0.39.1
    # via numba
matplotlib==3.6.2
    # via stegaplots (pyproject.toml)
numba==0.56.4
    # via stegaplots (pyproject.toml)
numpy==1.23.4
    # via
    #   contourpy
    #   matplotlib
    #   numba
    #   stegaplots (pyproject.toml)
packaging==21.3
    # via matplotlib
pillow==9.3.0
    # via matplotlib
pyparsing==3.0.9
    # via
    #   matplotlib
    #   packaging
python-dateutil==2.8.2
    # via matplotlib
six==1.16.0
    # via python-dateutil

# The following packages are considered to be unsafe in a requirements file:
# setuptools
